+++ sdk/cosmos/azure-cosmos/test/test_user_configs.py
@@ -22,15 +22,16 @@
 import unittest
 
 import azure.cosmos.cosmos_client as cosmos_client
+from azure.cosmos import http_constants
 import pytest
 from test_config import _test_config
-
 
 # This test class serves to test user-configurable options and verify they are
 # properly set and saved into the different object instances that use these
 # user-configurable settings.
 
 pytestmark = pytest.mark.cosmosEmulator
+
 
 @pytest.mark.usefixtures("teardown")
 class TestUserConfigs(unittest.TestCase):
@@ -45,6 +46,28 @@ class TestUserConfigs(unittest.TestCase):
         self.assertFalse(client_false.client_connection.connection_policy.EnableEndpointDiscovery)
         self.assertTrue(client_default.client_connection.connection_policy.EnableEndpointDiscovery)
         self.assertTrue(client_true.client_connection.connection_policy.EnableEndpointDiscovery)
+
+    def test_default_account_consistency(self):
+        # These tests use the emulator, which has a default consistency of "Session"
+        # If your account has a different level of consistency, make sure it's not the same as the custom_level below
+
+        client = cosmos_client.CosmosClient(url=_test_config.host, credential=_test_config.masterKey)
+        database_account = client.get_database_account()
+        account_consistency_level = database_account.ConsistencyPolicy["defaultConsistencyLevel"]
+        self.assertEqual(client.client_connection.default_headers[http_constants.HttpHeaders.ConsistencyLevel],
+                         account_consistency_level)
+
+        # Now testing a user-defined consistency level as opposed to using the account one
+        custom_level = "Strong"
+        client = cosmos_client.CosmosClient(url=_test_config.host, credential=_test_config.masterKey,
+                                            consistency_level=custom_level)
+        database_account = client.get_database_account()
+        account_consistency_level = database_account.ConsistencyPolicy["defaultConsistencyLevel"]
+        # Here they're not equal, since the headers being used make the client use a different level of consistency
+        self.assertNotEqual(
+            client.client_connection.default_headers[http_constants.HttpHeaders.ConsistencyLevel],
+            account_consistency_level)
+
 
 if __name__ == "__main__":
     unittest.main()